Fix implicit floating point precision loss

+++ Butter/BTRTextField.m
@@ -535,11 +535,11 @@ static void BTRCommonInit(BTRTextField *textField) {
 		NSSize textSize = [self cellSizeForBounds:theRect];
 		
 		// Center that in the proposed rect
-		float heightDelta = newRect.size.height - textSize.height;
+		CGFloat heightDelta = newRect.size.height - textSize.height;
 		if (heightDelta > 0)
 		{
 			newRect.size.height -= heightDelta;
-			newRect.origin.y += ceilf(heightDelta / 2);
+			newRect.origin.y += ceil(heightDelta / 2);
 		}
 	}
 	return [(BTRTextField *)[self controlView] drawingRectForProposedDrawingRect:newRect];
